Allow searching for articles (#62)
Articles are intended to be indexed in Elasticsearch. A Docker Compose configuration is supplied to manage the dependecy on Elasticsearch, but in theory the search backend could be a relational database. The AUTO_UPDATE option is enabled, as is ATOMIC_REBUILD. Together these ensure that the search index is updated whenever an article is changed, however, while this is convenient for development, it may not be appropriate for production.