Implement pages that are linked from the site footer (#96)
This has been re-implemented as part of Wagtail's Page hierarchy so that Footer Pages can be subject to access control restrictions. This was necessary for the multi-tenant approach where all sites are served from one Wagtail instance, but arguably this is not required now since a decision has been made to stick to single-tenancy. In any case, the approach is similar to v1, so should be familiar to site admins. One notable difference from v1 is that footer pages are listed by (treebeard) path by default, instead of title.


Co-authored-by: Ian Stride <ianstride@gmail.com>