Set up cache backend only if explicitly specified
Previously, if no cache was specified a 'dummy' cache would be used by default.
The dummy cache will not pass the healthcheck, and therefore the app will not
work. To fix this I have opted to remove the cache config entirely if no cache
is explicitly specified via the 'CACHE_BACKEND' env var.