Track additional site ID with Matomo (#1645)
The purpose of this feature is to allow tracking of the same site to two different site IDs in Matomo. This is deemed necessary to emulate what was occurring with Google Analytics before the migration to Matomo. Each site is tracked individually via its own site ID and in aggregate with other sites via a 'global' site ID. The global site ID tracks metrics from all sites to allow an overview for IoGT project management, whereas the individual site ID is for UNICEF country offices to view the metrics from their own site only.