Create session-based user tracking ID
When JavaScript is disabled in the browser, we use the Matomo Image Tracker to
gather analytics data. In this case, first party cookies are also disabled and
Matomo is not able to differentiate between different visitors very well. The
tracking ID is intended to provide the same function as the visitor ID that is
usually provided by Matomo when JS is enabled.