Generate a visitor ID instead of a user ID
The user ID is supposed to identify a registered user of a site, but we are
actually using it to identify a visitor. This change generates a visitor ID in a
similar way to the Matomo JS client, for the particular case where the visitor
does not have a JS-enabled browser. The user ID feature can then be implemented
later on without coming into conflict with this visitor ID feature.